--- coalesced_derive/tests/named_enum.rs
@@ -1,17 +1,0 @@
-use coalesced::Coalesce;
-
-#[derive(Coalesce)]
-enum Config {
-    Unit,
-    Named { value: i32 },
-    Unnamed(String),
-}
-
-#[test]
-fn test_derive_named_fields_struct() {
-    let config = Config::Unit;
-    let config2 = Config::Unit;
-
-    let c = config.prior(config2);
-    assert!(matches!(c, Config::Unit));
-}
