+++ coalesced_derive/tests/named_fields_struct_generics.rs
@@ -7,7 +7,7 @@ struct Config<T> {
 }
 
 #[test]
-fn test_derive_generics_named_fields_struct() {
+fn test_derive_named_fields_struct_generics() {
     let config = Config {
         name: "c1",
         value: Some(1),
