+++ coalesced_derive/tests/unnamed_fields_struct_constraint_generics.rs
@@ -8,7 +8,7 @@ where
     S: Display + Coalesce;
 
 #[test]
-fn test_derive_named_fields_struct() {
+fn test_derive_unnamed_fields_struct_constraint_generics() {
     let config = Config("c1", Some(1));
     let config2 = Config("c2", None);
